+++ shift-scan/src/app/(routes)/admins/inbox/@view/[id]/_components/inboxHeader.tsx
@@ -9,7 +9,7 @@ import { TextAreas } from "@/components/(reusable)/textareas";
 import { Texts } from "@/components/(reusable)/texts";
 import { useTranslations } from "next-intl";
 import { Dispatch, SetStateAction, useState } from "react";
-import { LeaveRequest } from "../page";
+import { LeaveRequest } from "@/lib/types";
 
 export function RequestHeader({
   leaveRequest,
